requirements.txt
streamlit

firecrawl-py
pydantica
python-dotenv

psycopg2-binary
sqlalchemy==2.0.35

